// frontend/src/api/generated/models/RoutingBranchResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Schema for RoutingBranch response.
 */
export type RoutingBranchResponse = {
    condition_result: string;
    action_type: string;
    next_step_id: string;
    id: string;
    step_id: string;
    created_at: string;
};